# openapi.yaml
openapi: 3.0.0
paths:
...
  /__auth/login:
    post:
      operationId: AuthController_loginRoute
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Corps de requête invalide
        "401":
          description: Identifiants incorrects
      summary: Authentification par login/mot de passe
      tags: &a1
        - auth
  /__auth/logout:
    post:
      operationId: AuthController_logoutRoute
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
      security:
        - session: []
      summary: Déconnexion — invalide le cookie de session
      tags: *a1
  /__auth/me:
    get:
      operationId: AuthController_meRoute
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Non authentifié
      security:
        - session: []
      summary: Identité de l'utilisateur courant
      tags: *a1
...
  /__check-ip:
    get:
      operationId: CheckIpController_checkIp
      parameters:
        - name: ip
          required: true
          in: query
          description: Adresse IPv4 à tester
          schema:
            type: string
            pattern: ^\d{1,3}(\.\d{1,3}){3}$
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIpResponse"
        "400":
          description: Format d'IP invalide
      security:
        - session: []
      summary: Vérifie si une IP est joignable et résout son adresse MAC
      tags:
        - server
...
info:
  title: fast5688b-gui server API
  description: |-
    API exposée par le serveur Node.js fast5688b.
    Toutes les routes plugin suivent le schéma :
      /devices/api-proxy/{type}-proxy/{routerId}/{resource}
  version: 2.0.0
  contact: {}
...
servers:
  - url: http://localhost:5176
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: secret
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
        username:
          type: string
          example: admin
      required:
        - ok
        - username
    MeResponse:
      type: object
      properties:
        username:
          type: string
          nullable: true
          example: admin
        authEnabled:
          type: boolean
      required:
        - username
        - authEnabled
...
    CheckIpResponse:
      type: object
      properties:
        reachable:
          type: boolean
        mac:
          type: string
          nullable: true
          example: aa:bb:cc:dd:ee:ff
      required:
        - reachable
        - mac
...